Time: 0 ms (100%), Space: 7.9 MB (62%) - LeetHub

+++ 0096-unique-binary-search-trees/0096-unique-binary-search-trees.cpp
@@ -5,8 +5,8 @@ public:
         // Catalan number: C(n) = (2n)! / (n+1)! * n!
         // Use tgamma(n+1) == n!
 
-        long long numerator = tgamma(2*n + 1);     
-        long long denominator = tgamma(n + 2) * tgamma(n + 1);  
+        double numerator = tgamma(2*n + 1);     
+        double denominator = tgamma(n + 2) * tgamma(n + 1);  
 
         return static_cast<int>(numerator / denominator + 0.5);  // round to nearest int
     }
